% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cty_cz_2020_xwalk.R
\docType{data}
\name{cty_cz_2020_xwalk}
\alias{cty_cz_2020_xwalk}
\title{Relationship file (crosswalk) between US counties and commuting zones}
\format{
\subsection{A data frame with 2 columns and 3222 rows:}{

\describe{
\item{GEOID}{US county FIPS code.}
\item{GEOID2}{Commuting zone ID for the year 2020.}

}
}
}
\source{
https://sites.psu.edu/psucz/data/
}
\usage{
cty_cz_2020_xwalk
}
\description{
This table contains GEOIDs for US counties and the commuting zones they are
nested in. Commuting zones can be used to study regional economy with
considerations of urban-rural interconnections across state lines. For
details refer to papers by  \href{https://doi.org/10.1007/s11113-016-9386-0}{Fowler Jensen and Rhubart (2016)} and \href{https://doi.org/10.17605/OSF.IO/J256U}{Fowler (2024)}.
}
\keyword{datasets}
