% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSVI-package.R
\docType{package}
\name{findSVI-package}
\alias{findSVI}
\alias{findSVI-package}
\title{findSVI: Calculate CDC/ATSDR Social Vulnerability Index}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Retrieve variables from US census data and calculate SVI in a given geographic area based on CDC/ATSDR SVI documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heli-xu/findSVI}
  \item \url{https://heli-xu.github.io/findSVI/}
  \item Report bugs at \url{https://github.com/heli-xu/findSVI/issues}
}

}
\author{
\strong{Maintainer}: Heli Xu \email{xuheli91@gmail.com}

}
\keyword{internal}
