% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_data.R
\name{get_census_data}
\alias{get_census_data}
\title{Retrieve census data for American Community Survey (with Tidycensus)}
\usage{
get_census_data(
  year,
  state = NULL,
  geography,
  county = NULL,
  key = NULL,
  geometry = FALSE,
  ...
)
}
\arguments{
\item{year}{The year of interest (available 2012-2021).}

\item{state}{(Optional) Specify the state of interest. If data for multiple
states are retrieved together, ranking for SVI calculation will be performed
among all states. `state = NULL` as default, or `state = 'US'` return
nation-level data.}

\item{geography}{The geography of interest (eg. state, county, zcta, tract)}

\item{county}{(Optional) Specify the county(s) of interest, must be combined
with a value supplied to "state".}

\item{key}{Your Census API key. obtain one at
<https://api.census.gov/data/key_signup.html>.}

\item{geometry}{Default as FALSE for a regular tibble of census data. If set
as TRUE, returns a tibble with an additional `geometry` column containing
simple feature geometry.}

\item{...}{Other arguments; more details please see [tidycensus::get_acs()]}
}
\value{
A tibble of ACS data with each row represents an enumeration unit and
 each column represents the variables ("wide" form).
}
\description{
This function uses [tidycensus::get_acs()] with a pre-defined
 list of variables to retrieves ACS data for SVI calculation. Note that a
 Census API key is required for this function to work, which can be obtained
 at <https://api.census.gov/data/key_signup.html>. To set up your key, use
 `census_api_key("YOUR KEY GOES HERE")`, or include it as an argument.
}
\examples{
\dontrun{
census_api_key("YOUR KEY GOES HERE")
get_census_data(2018, "county", "PA")}
}
