% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSVI-package.R
\docType{package}
\name{findSVI-package}
\alias{findSVI}
\alias{findSVI-package}
\title{findSVI: Calculate Social Vulnerability Index for Communities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Developed by CDC/ATSDR (Centers for Disease Control and Prevention/ Agency for Toxic Substances and Disease Registry), Social Vulnerability Index (SVI) serves as a tool to assess the resilience of communities by taking into account socioeconomic and demographic factors. Provided with year(s), region(s) and a geographic level of interest, 'findSVI' retrieves required variables from US census data and calculates SVI for communities in the specified area based on CDC/ATSDR SVI documentation. Reference for the calculation methods: Flanagan BE, Gregory EW, Hallisey EJ, Heitgerd JL, Lewis B (2011) \doi{10.2202/1547-7355.1792}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heli-xu/findSVI}
  \item \url{https://heli-xu.github.io/findSVI/}
  \item Report bugs at \url{https://github.com/heli-xu/findSVI/issues}
}

}
\author{
\strong{Maintainer}: Heli Xu \email{xuheli91@gmail.com} (\href{https://orcid.org/0000-0002-9792-2727}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ran Li \email{rl627@drexel.edu} (\href{https://orcid.org/0000-0002-4699-4755}{ORCID})
  \item Usama Bilal \email{ub45@drexel.edu} (\href{https://orcid.org/0000-0002-9868-7773}{ORCID})
}

}
\keyword{internal}
